refactor(next): move field count api to `/ticket-fields/count`
