+++ next/api/src/common/pipe/FindModelPipe.ts
@@ -37,3 +37,22 @@ export class FindModelWithoutDeleteFlagPipe<M extends typeof Model> {
     return instance;
   }
 }
+
+export class FindModelOptionalPipe<M extends typeof Model> {
+  constructor(private model: M, private authOptions?: AuthOptions) {}
+
+  async transform(id: string | undefined, ctx: Context): Promise<InstanceType<M> | undefined> {
+    if (!id) {
+      return undefined;
+    }
+
+    const currentUser = ctx.state.currentUser as User | undefined;
+
+    const instance = await this.model.find(id, this.authOptions ?? currentUser?.getAuthOptions());
+    if (!instance) {
+      throw new NotFoundError(`${this.model.getClassName()} "${id}"`);
+    }
+
+    return instance;
+  }
+}
