+++ next/api/src/controller/category.ts
@@ -1,7 +1,5 @@
 import { Context } from 'koa';
 import { z } from 'zod';
-import axios from 'axios';
-import { escape } from 'sqlstring';
 
 import {
   Body,
@@ -9,7 +7,6 @@ import {
   Ctx,
   CurrentUser,
   Get,
-  InternalServerError,
   Param,
   Patch,
   Post,
@@ -25,13 +22,11 @@ import { Category } from '@/model/Category';
 import { TicketForm } from '@/model/TicketForm';
 import { User } from '@/model/User';
 import { ArticleAbstractResponse } from '@/response/article';
-import { CategoryFieldStatsResponse, CategoryResponse } from '@/response/category';
+import { CategoryResponse } from '@/response/category';
 import { CategoryFieldResponse } from '@/response/ticket-field';
 import { ArticleTopicFullResponse } from '@/response/article-topic';
 import { getTopic } from '@/model/ArticleTopic';
 import _ from 'lodash';
-import { OPTION_TYPES } from '@/model/TicketField';
-import { Status } from '@/model/Ticket';
 import { FindCategoryPipe, categoryService } from '@/category';
 
 const createCategorySchema = z.object({
@@ -201,148 +196,6 @@ export class CategoryController {
     @Query('active', new ParseBoolPipe({ keepUndefined: true })) active?: boolean
   ) {
     return await categoryService.getSubCategories(category.id, active);
-  }
-
-  @Get(':id/count')
-  @UseMiddlewares(auth, customerServiceOnly)
-  async getFieldCounts(
-    @Param('id', FindCategoryPipe) category: Category,
-    @Ctx() ctx: Context,
-    @Query('from') from?: string,
-    @Query('to') to?: string
-  ) {
-    const optionFields = await (await category.load('form', { useMasterKey: true }))?.load(
-      'fields',
-      { useMasterKey: true }
-    );
-
-    const localesMap = new Map(ctx.locales?.map((locale, index) => [locale, index + 1]));
-
-    if (!optionFields) {
-      return [];
-    }
-
-    const skipSerialize = Symbol('skipSerialize');
-
-    return (
-      await Promise.allSettled(
-        optionFields.map(async (optionField) => {
-          const variants = await optionField.load('variants', { useMasterKey: true });
-
-          // skip process if variant is not option type or doesn't exist
-          if (!variants || !OPTION_TYPES.includes(optionField.type)) {
-            return Promise.reject(skipSerialize);
-          }
-
-          const fieldValues = [
-            ...variants.reduce<Map<string, [string, string]>>((pre, variant) => {
-              variant.options?.forEach((option) => {
-                const optionInStore = pre.get(option.value);
-
-                if (optionInStore) {
-                  const [, locale] = optionInStore;
-                  const curOptionPriority = localesMap.get(variant.locale);
-                  const optionInStorePriority = localesMap.get(locale);
-
-                  if (
-                    // override if cur option has priority but stored is not
-                    (!optionInStorePriority && curOptionPriority) ||
-                    // override if cur option has higher priority
-                    (curOptionPriority &&
-                      optionInStorePriority &&
-                      curOptionPriority < optionInStorePriority) ||
-                    // override if stored option has no priority and cur option's locale is default
-                    (!optionInStorePriority && variant.locale === optionField.defaultLocale)
-                  ) {
-                    pre.set(option.value, [option.title, variant.locale]);
-                  }
-
-                  return;
-                }
-
-                pre.set(option.value, [option.title, variant.locale]);
-              });
-              return pre;
-            }, new Map()),
-          ];
-
-          const getSql = (fieldId: string, fieldValue: string, categoryId: string) => `
-            SELECT count(t.objectId) AS count, t.status
-            FROM Ticket AS t
-            LEFT JOIN TicketFieldValue AS v
-            ON t.objectId = v.\`ticket.objectId\`
-            WHERE arrayExists(
-                x -> visitParamExtractString(x, 'field') = ${escape(fieldId)}
-                AND (
-                    visitParamExtractString(x, 'value') = ${escape(fieldValue)}
-                    OR
-                    arrayExists(x -> x = ${escape(
-                      fieldValue
-                    )}, JSONExtract(x, 'value', 'Array(String)'))
-                ),
-                v.values
-            )
-            AND
-            visitParamExtractString(t.category, 'objectId') = ${escape(categoryId)}
-            ${(from && `AND t.createdAt >= parseDateTimeBestEffortOrNull(${escape(from)})`) || ''}
-            ${(to && `AND t.createdAt <= parseDateTimeBestEffortOrNull(${escape(to)})`) || ''}
-            GROUP BY t.status
-          `;
-
-          const options = await Promise.all(
-            fieldValues.map(async ([fieldValue, [fieldTitle, titleLocale]]) => {
-              const res = await axios.get<{
-                results: { count: number; status: number }[];
-              }>(`${process.env.LC_API_SERVER as string}/datalake/v1/query`, {
-                params: { sql: getSql(optionField.id, fieldValue, category.id) },
-                headers: {
-                  'X-LC-ID': process.env.LC_APP_ID as string,
-                  'X-LC-Key': (process.env.LC_APP_MASTER_KEY as string) + ',master',
-                },
-              });
-
-              return {
-                title: fieldTitle,
-                displayLocale: titleLocale,
-                value: fieldValue,
-                count: res.data.results.reduce(
-                  (acc, { count, status }) =>
-                    Status.isOpen(status)
-                      ? { ...acc, open: acc.open + Number(count), total: acc.total + Number(count) }
-                      : {
-                          ...acc,
-                          closed: acc.closed + Number(count),
-                          total: acc.total + Number(count),
-                        },
-                  { open: 0, closed: 0, total: 0 }
-                ),
-              };
-            }, [])
-          );
-          return {
-            id: optionField.id,
-            title: optionField.title,
-            type: optionField.type,
-            options: options.sort(
-              ({ count: { total: totalA } }, { count: { total: totalB } }) => totalB - totalA
-            ),
-          };
-        })
-      )
-    )
-      .reduce<CategoryFieldStatsResponse>((acc, field) => {
-        if (field.status === 'fulfilled') {
-          return [...acc, field.value];
-        }
-        // remove skipped fields from this array
-        if (field.reason !== skipSerialize) {
-          throw new InternalServerError(field.reason as string);
-        }
-        return acc;
-      }, [])
-      .sort(({ title: titleA }, { title: titleB }) =>
-        titleA < titleB ? -1 : titleA === titleB ? 0 : 1
-      );
   }
 
   private convertUpdateData(data: UpdateCategoryData): UpdateData<Category> {
