+++ next/web/src/App/Admin/Settings/index.tsx
@@ -23,7 +23,7 @@ import TimeTriggerDetail from './Automations/TimeTriggers/Detail';
 import { ArticleRevisionDetail, ArticleRevisions } from './Articles/Revision';
 import { Weekday } from './Others';
 import { NewTopic, TopicDetail, TopicList } from './Topics';
-import { CategoryFieldStats } from './Categories/CategoryFieldStats';
+import { CategoryFieldCount } from './Categories/CategoryFieldStats';
 
 const SettingRoutes = () => (
   <Routes>
@@ -37,7 +37,7 @@ const SettingRoutes = () => (
     <Route path="/categories">
       <Route index element={<CategoryList />} />
       <Route path="new" element={<NewCategory />} />
-      <Route path="count/:id" element={<CategoryFieldStats />} />
+      <Route path="count/:id" element={<CategoryFieldCount />} />
       <Route path=":id" element={<CategoryDetail />} />
     </Route>
     <Route path="/tags">
